Fix code for word 'Qat'
The word 'Qat' does not have the part-of-speech, frequency or usage
information. Also, the word-separation property is in different block
than usually (in the main <section> rather than the 'Rechtschreibung'
<section>).

Also, beautifulSoup 4.4 might be needed beacuse of the usage of
copy.copy() function. The with help of the copy function parts of the
html DOM are cloned so that temporary deletions of nodes do not affect
the original DOM.