Add possibility to load word from soup object
If caching is in place, making a http request every time a word is
loaded is not neccessary.

This commit introduces the duden.load_soup function which takes a
BeautifulSoup object and returns a DudenWord object. This is in
contrary to the duden.get function which makes a http GET request
every time.

Fixes #20.