Remove unused 'exists' property
The exists property was used when DudenWord objects were still
initialized using the word's urlname. If the page failed to load (error
code 404) then the DudenWord object was still created, but had the
exists property returning False.

This was changed with the introduction of the load_soup function when
also the behaviour of the DudenWord constructor changed - it takes the
BeautifulSoup object (created from the corresponding duden.de page) - so
the page always exists when the word object is created.