Add possiblity to not resolve words during search
If duden.search finds 10 word matches, it will try to create 10
DudenWord objects, which means to make 10 http GET requests. This is not
always desirable.

The return_words=False of duden.search function will cause it to return
only the url names, so for example, searching for 'einfach' would return
['einfach_vollkommen_wirklich', 'einfach_einmal_simpel']
.