Search for words instead of word url representations
Every word in duden database has a corresponding url - for example - for
the word müßig it's
https://www.duden.de/rechtschreibung/mueszig .

Previously, one had to type in the url form, that is mueszig. With this
commit, users can use the word itself (with all accents) to search.

This commit adds two CLI arguments:
* -r/--result to choose the correct word in case of multiple matching
words
* --fuzzy to enable fuzzy searching (searches for parts of words, etc.)

Fixes #44.